+++ src/Rik.StatusPage.AspNetCore/StatusPageOptions.cs
@@ -1,0 +1,9 @@
+using Microsoft.AspNetCore.Http;
+
+namespace Rik.StatusPage.AspNetCore
+{
+    public class StatusPageOptions
+    {
+        public PathString Path { get; set; } = "/status.xml";
+    }
+}