+++ src/Rik.StatusPage/Providers/CustomStatusProvider.cs
@@ -1,30 +1,33 @@
 ﻿using System;
+using System.Threading.Tasks;
 using Rik.StatusPage.Configuration;
+using Rik.StatusPage.Internal;
 using Rik.StatusPage.Schema;
 
 namespace Rik.StatusPage.Providers
 {
-    public class CustomStatusProvider : StatusProvider
+    public class CustomStatusProvider : StatusProvider<CustomStatusProviderOptions>
     {
-        private readonly StatusProvider customStatusProvider;
+        private readonly StatusProvider<CustomStatusProviderOptions> customStatusProvider;
 
-        public CustomStatusProvider(StatusProviderConfigurationElement configuration)
-            : base(configuration)
+        public override string DisplayUri => customStatusProvider.DisplayUri;
+
+        public CustomStatusProvider(CustomStatusProviderOptions options)
+            : base(options)
         {
-            if (string.IsNullOrWhiteSpace(configuration.Class))
-                throw new ArgumentException("Class name is required for custom provider.", nameof(configuration.Class));
+            if (string.IsNullOrWhiteSpace(options.Class))
+                throw new ArgumentException("Class name is required for custom provider.", nameof(options.Class));
 
-            var statusProviderType = FindType(configuration.Class);
-
+            var statusProviderType = TypeHelper.FindType(options.Class);
             if (statusProviderType == null)
-                throw new Exception($"Cannot load custom provider type: {configuration.Class}.");
+                throw new Exception($"Cannot load custom provider type: {options.Class}.");
 
-            customStatusProvider = (StatusProvider)Activator.CreateInstance(statusProviderType, configuration);
+            customStatusProvider = (StatusProvider<CustomStatusProviderOptions>)Activator.CreateInstance(statusProviderType, options);
         }
 
-        protected override ExternalUnit OnCheckStatus(ExternalUnit externalUnit)
+        protected override Task<ExternalUnit> OnCheckStatusAsync(ExternalUnit externalUnit)
         {
-            return customStatusProvider.CheckStatus();
+            return customStatusProvider.CheckStatusAsync();
         }
     }
 }