+++ src/Shared/Configuration/DatabaseStatusProviderOptions.cs
@@ -1,0 +1,7 @@
+namespace Rik.StatusPage.Configuration
+{
+    public class DatabaseStatusProviderOptions : StatusProviderOptions
+    {
+        public string ConnectionString { get; set; }
+    }
+}