+++ src/Shared/Configuration/FileStorageStatusProviderOptions.cs
@@ -1,0 +1,9 @@
+namespace Rik.StatusPage.Configuration
+{
+    public class FileStorageStatusProviderOptions : StatusProviderOptions
+    {
+        public string StoragePath { get; set; }
+        public bool RequireRead { get; set; }
+        public bool RequireWrite { get; set; }
+    }
+}