+++ src/Shared/Configuration/StatusProviderOptions.cs
@@ -1,0 +1,8 @@
+namespace Rik.StatusPage.Configuration
+{
+    public class StatusProviderOptions
+    {
+        public string Name { get; set; }
+        public string DisplayUri { get; set; }
+    }
+}