+++ src/Shared/Configuration/WebServiceStatusProviderOptions.cs
@@ -1,0 +1,7 @@
+namespace Rik.StatusPage.Configuration
+{
+    public class WebServiceStatusProviderOptions : StatusProviderOptions
+    {
+        public string Uri { get; set; }
+    }
+}