+++ src/Shared/Providers/IStatusProvider.cs
@@ -1,0 +1,10 @@
+using System.Threading.Tasks;
+using Rik.StatusPage.Schema;
+
+namespace Rik.StatusPage.Providers
+{
+    public interface IStatusProvider
+    {
+        Task<ExternalUnit> CheckStatusAsync();
+    }
+}