+++ src/Shared/Providers/OracleDatabaseStatusProvider.cs
@@ -14,8 +14,10 @@ namespace Rik.StatusPage.Providers
         protected override string VersionQuery => "SELECT version FROM V$INSTANCE";
         protected override string PlatformName { get; } = "Oracle Database";
 
-        public OracleDatabaseStatusProvider(StatusProviderConfigurationElement configuration)
-            : base(configuration)
+        public override string DisplayUri => "";
+
+        public OracleDatabaseStatusProvider(DatabaseStatusProviderOptions options)
+            : base(options)
         { }
     }
 }