+++ RELEASE_NOTES.md
@@ -1,4 +1,11 @@
-﻿#### 1.0.1 - 13.11.2017
+﻿#### 1.1.0 - Unreleased
+* New key used to sign the assembly
+* Status page middleware for AspNetCore applications
+* Use asynchronous workflow
+* Upgrade RabbitMq provider to EasyNetQ 1.0.0
+* Improved provider options (strong types).
+
+#### 1.0.1 - 13.11.2017
 * RabbitMQ status provider
 * Uri value from configuration
 * Version value for app from configuration